+++ src/libs/ajax.js
@@ -1,8 +1,7 @@
-import { getDefaultProperties } from '@databiosphere/bard-client'
 import _ from 'lodash/fp'
 import * as qs from 'qs'
 import {
-  appIdentifier, authOpts, checkRequesterPaysError, fetchAgora, fetchBard, fetchBillingProfileManager, fetchBond, fetchCatalog,
+  appIdentifier, authOpts, checkRequesterPaysError, fetchAgora, fetchBillingProfileManager, fetchBond, fetchCatalog,
   fetchDataRepo, fetchDockstore,
   fetchDrsHub,
   fetchEcm, fetchGoogleForms,
@@ -12,14 +11,13 @@ import { Apps } from 'src/libs/ajax/Apps'
 import { AzureStorage } from 'src/libs/ajax/AzureStorage'
 import { Disks } from 'src/libs/ajax/Disks'
 import { GoogleStorage } from 'src/libs/ajax/GoogleStorage'
+import { Metrics } from 'src/libs/ajax/Metrics'
 import { Runtimes } from 'src/libs/ajax/Runtimes'
-import { ensureAuthSettled, getUser } from 'src/libs/auth'
+import { getUser } from 'src/libs/auth'
 import { getConfig } from 'src/libs/config'
 import { withErrorIgnoring } from 'src/libs/error'
-import * as Nav from 'src/libs/nav'
-import { authStore, knownBucketRequesterPaysStatuses, requesterPaysProjectStore, userStatus, workspaceStore } from 'src/libs/state'
+import { knownBucketRequesterPaysStatuses, requesterPaysProjectStore, workspaceStore } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
-import { v4 as uuid } from 'uuid'
 
 
 window.ajaxOverrideUtils = {
@@ -1211,41 +1209,6 @@ const Duos = signal => ({
     const res = await fetchOrchestration(`/api/duos/consent/orsp/${orspId}`, _.merge(authOpts(), { signal }))
     return res.json()
   }
-})
-
-const Metrics = signal => ({
-  captureEvent: withErrorIgnoring(async (event, details = {}) => {
-    await ensureAuthSettled()
-    const { isSignedIn, registrationStatus } = authStore.get() // NOTE: This is intentionally read after ensureAuthSettled
-    const isRegistered = isSignedIn && registrationStatus === userStatus.registeredWithTos
-    if (!isRegistered) {
-      authStore.update(_.update('anonymousId', id => {
-        return id || uuid()
-      }))
-    }
-    const body = {
-      event,
-      properties: {
-        ...details,
-        distinct_id: isRegistered ? undefined : authStore.get().anonymousId,
-        appId: 'Saturn',
-        hostname: window.location.hostname,
-        appPath: Nav.getCurrentRoute().name,
-        ...getDefaultProperties()
-      }
-    }
-
-    return fetchBard('api/event', _.mergeAll([isRegistered ? authOpts() : undefined, jsonBody(body), { signal, method: 'POST' }]))
-  }),
-
-  syncProfile: withErrorIgnoring(() => {
-    return fetchBard('api/syncProfile', _.merge(authOpts(), { signal, method: 'POST' }))
-  }),
-
-  identify: withErrorIgnoring(anonId => {
-    const body = { anonId }
-    return fetchBard('api/identify', _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }]))
-  })
 })
 
 const OAuth2 = signal => ({
